clean out sources and analysis
